MCP/requirements.txt
astropy
keras
lxml
numpy
opencv-python
pillow
pyindi-client
pyserial
pysolar
pytz
requests
scikit-image
scipy
shapely
tensorflow
